Fixing ctrl-v and lines numbers

+++ src/context/ui/elements/CodingTabArea.java
@@ -16,76 +16,67 @@ import mindustry.ui.Styles;
 import mindustry.ui.dialogs.BaseDialog;
 
 public class CodingTabArea extends TabArea {
-    public String code = "";
+    private final ColoredTextArea codeArea = new ColoredTextArea("");
+    private final Label lineLabel = new Label("") {
+        @Override
+        public float getPrefHeight() {
+            return 0f;
+        }
+    };
 
-    private Label lineLabel;
-    private TextArea codeArea;
-    private Slider slider;
-
+    /**
+     * Table that is behind the codeArea, used to set the backColor of the codeArea
+     */
+    private final Table behindTable = new Table();
 
     private TextButton buttonSync;
     private Fi synchronizedFile = null;
-    public Cons<Fi> onSynchronize = (Fi) -> {
-    };
-    public long lastFileTime = 0;
-    public Timer.Task verifyFile = null;
+    private Cons<Fi> onSynchronize = fi -> {};
+    private long lastFileTime = 0;
+    private Timer.Task verifyFile = null;
 
 
     public CodingTabArea() {
         super();
-    }
-
-    private void updateLineHeight() {
-        StringBuilder txt = new StringBuilder();
-        int minimum = codeArea.getFirstLineShowing();
-        int maximum = codeArea.getFirstLineShowing() + codeArea.getLinesShowing();
-        for (int i = minimum; i < maximum - 1; i++) {
-            String sendString = String.valueOf(i + 1);
-            if (sendString.length() < String.valueOf(maximum).length()) txt.append(" ");
-            txt.append(sendString).append("\n");
-        }
-        lineLabel.invalidate();
-        lineLabel.setText(txt.toString());
-
-        if (codeArea.getLines() > codeArea.getLinesShowing()) {
-            slider.setRange(0, codeArea.getLines() - codeArea.getLinesShowing());
-            slider.setValue(codeArea.getFirstLineShowing());
-        }
-    }
-
-    @Override
-    public void generateBase(Table table) {
-        super.generateBase(table);
-
         //line numbers
         Label.LabelStyle labelStyle = new Label.LabelStyle(Styles.outlineLabel);
         labelStyle.background = Styles.black8;
-        lineLabel = new Label("1\n2\n3\n4\n5\n6\n7\n8\n9\n10\n11\n12\n13\n14\n15\n16\n17\n18\n19\n20\n21\n22\n23\n24\n25\n26\n27\n28\n29\n30\n31\n32\n33\n34\n35\n36\n37");
+
+        lineLabel.setText(() -> {
+            int firstLine = codeArea.getFirstLineShowing();
+            StringBuilder sb = new StringBuilder();
+
+            int cursorLine = codeArea.getCursorLine();
+
+            for (int i = 0; i < codeArea.getLinesShowing(); i++) {
+                if (i + firstLine == cursorLine) sb.append("[yellow]");
+
+                sb.append(firstLine + i + 1).append("\n");
+
+                if (i + firstLine == cursorLine) sb.append("[]");
+            }
+            return sb.toString();
+        });
         lineLabel.setStyle(labelStyle);
         lineLabel.setAlignment(Align.topRight);
 
         // Code backColor
-        Table behindTable = new Table();
         behindTable.setBackground(Styles.grayPanel);
 
         // Code Area
         TextField.TextFieldStyle codeAreaStyle = new TextField.TextFieldStyle(Styles.defaultField);
         codeAreaStyle.background = null;
-        codeArea = new ColoredTextArea(code);
         codeArea.setStyle(codeAreaStyle);
-        codeArea.setTextFieldListener((textField, c) -> {
-            code = textField.getText();
-            updateLineHeight();
-        });
+        codeArea.changed(() -> setCode(codeArea.getText()));
 
-        // ScrollBar
-        slider = new Slider(0, 1, 1f, false);
-        slider.visible(() -> this.codeArea != null && this.codeArea.getLinesShowing() < this.codeArea.getLines());
+        behindTable.add(codeArea).grow();
+    }
 
-        // Adding to Table
+    @Override
+    public void generateBase(Table table) {
+        super.generateBase(table);
         table.add(lineLabel).left().top().padRight(5);
         table.add(behindTable).grow();
-        behindTable.add(codeArea).grow();
     }
 
     @Override
@@ -99,16 +90,20 @@ public class CodingTabArea extends TabArea {
             }
             Vars.platform.showFileChooser(false, "js", fi -> {
                 String content = fi.readString();
-                if(content.equals(this.code)) {
-                    this.setSync(fi,false);
+                if (content.equals(this.getCode())) {
+                    this.setSync(fi, false);
                 }
                 CodingTabArea cta = this;
-                this.syncTypeAsk(true, new ExpImpListener(){
+                this.syncTypeAsk(true, new ExpImpListener() {
                     @Override
-                    public void upload() { cta.setSync(fi, true);}
+                    public void upload() {
+                        cta.setSync(fi, true);
+                    }
 
                     @Override
-                    public void download() { cta.setSync(fi, false);}
+                    public void download() {
+                        cta.setSync(fi, false);
+                    }
 
                     @Override
                     public void delete() {
@@ -120,11 +115,10 @@ public class CodingTabArea extends TabArea {
         });
         Button searchTerm = new TextButton("@context.search-term");
         searchTerm.clicked(() -> {
-            SearchFunction sf = new SearchFunction();
             BaseDialog bd = new BaseDialog("Search Function");
-            bd.cont.add(sf.cont).grow();
+            bd.cont.add(SearchFunction.cont).grow();
             bd.cont.setBackground(Tex.button);
-            bd.buttons.button("@exit", bd::hide).size(120f,60f);
+            bd.buttons.button("@exit", bd::hide).size(120f, 60f);
             bd.closeOnBack();
             bd.show();
         });
@@ -135,7 +129,6 @@ public class CodingTabArea extends TabArea {
 
     public void setCode(String code) {
         this.codeArea.setText(code);
-        this.code = code;
     }
 
     public void setSync() {
@@ -145,8 +138,9 @@ public class CodingTabArea extends TabArea {
         if (verifyFile != null) verifyFile.cancel();
         verifyFile = null;
     }
+
     public void setSync(Fi file, boolean replaceFile) {
-        onSynchronize.get(file);
+        getOnSynchronize().get(file);
         this.synchronizedFile = file;
         if (file == null) {
             this.setSync();
@@ -166,29 +160,31 @@ public class CodingTabArea extends TabArea {
         BaseDialog d = new BaseDialog("@choose");
         d.cont.label(() -> "@context.code-ide.difference");
         d.closeOnBack();
-        d.buttons.button("@context.delete", Icon.trash, () -> {
-            cb.delete();
-            d.hide();
-        }).size(230f,60f);
+        if (resetOption) {
+            d.buttons.button("@context.delete", Icon.trash, () -> {
+                cb.delete();
+                d.hide();
+            }).size(230f, 60f);
+        }
         d.buttons.button("@context.upload", Icon.upload, () -> {
             cb.upload();
             d.hide();
-        }).size(230f,60f);
+        }).size(230f, 60f);
         d.buttons.button("@context.download", Icon.download, () -> {
             cb.download();
             d.hide();
-        }).size(230f,60f);
+        }).size(230f, 60f);
         d.buttons.button("@context.cancel", Icon.none, () -> {
             cb.cancel();
             d.hide();
-        }).size(230f,60f);
+        }).size(230f, 60f);
         d.show();
     }
 
     public void readFromFile(boolean replaceFile) {
         if (synchronizedFile == null) return;
         lastFileTime = synchronizedFile.lastModified();
-        if(replaceFile) synchronizedFile.writeString(this.code);
+        if (replaceFile) synchronizedFile.writeString(this.getCode());
         else this.setCode(synchronizedFile.readString());
         ide.blinkArea(Color.green, 1f);
     }
@@ -205,14 +201,23 @@ public class CodingTabArea extends TabArea {
         verifyFile = null;
     }
 
-    public static class ExpImpListener {
+    public String getCode() {
+        return this.codeArea.getText();
+    }
 
-        public ExpImpListener() {}
+    public Cons<Fi> getOnSynchronize() {
+        return onSynchronize;
+    }
 
-        public void upload(){}
-        public void download(){}
-        public void cancel(){}
-        public void delete(){}
+    public void setOnSynchronize(Cons<Fi> onSynchronize) {
+        this.onSynchronize = onSynchronize;
+    }
+
+    public abstract static class ExpImpListener {
+        public void upload() {}
+        public void download() {}
+        public void cancel() {}
+        public void delete() {}
     }
 
 }
